model.base.Referable: Add instantiation of backend to be called for update and commit
The backends.get_backend() function returns the Backend subclass. To
properly call the update_object()/commit_object() function, we need to
call them on an instance of this class. Therefore, add two brackets
behind each `get_backend()()` call like this, to immediatly get an
instance of the class. Note: This only works, if the class can be
initialized without setting parameters. While this should be the case,
it may be benefitial to notify the user of this constraint while writing
their own backend.