tests: Suppress MyPy warnings about TestCase.assertLogs
MyPy complains that the returned from the assertLogs() context manager
may be None. I assume this to be a MyPy or typeshed bug: The docs
clearly state that the 'recording helper' is always returned:
https://docs.python.org/3.10/library/unittest.html#unittest.TestCase.assertLogs

Thus, I have added 'type: ignore's here, so that MyPy can inform us (via
--warn-unused-ignores) when these are no longer needed -- in contrast to
'assert is not None' statements. As soon as this is the case, this
commit can be reverted.