test: set PYTHONPATH when running the compliance tool
By setting the PYTHONPATH when running the compliance tool in the tests,
it is no longer required to manually set PYTHONPATH when running the
tests with `python -m unittest` or `coverage`. Thus don't set PYTHONPATH
in the CI script anymore.
Furthermore abstract running the compliance tool with a function,
eliminating a lot of duplicate code.