requirements: update jsonschema to 4.7
This dependency is updated to fix a DeprecationWarning when running the unittests:

DeprecationWarning: The metaschema specified by $schema was not found.
Using the latest draft to validate, but this will raise an error in the future

This warning was logged because jsonschema didn't have support for
schema version 2019-09 in versions <4.