model.datatypes: add yearMonthDuration, dayTimeDuration, dateTimeStamp
These classes were missing from the implementation, because only XML
1.0 datatypes have been implemented so far:

*yearMonthDuration* is a restriction of duration, were only years
and months are allowed.

*dayTimeDuration* is a restriction of duration, were specifying
years and months is not allowed.

*dateTimeStamp* is a restriction of dateTime, were the timezone is
required.

yearMonthDuration and dayTimeDuration are implemented as subclasses of
Duration. The constraints are only checked on serialization, which makes
it a bit hacky. However, since this commit will be reverted with DotAAS
3.0 anyways, it's good enough for now.

Close #49