model: cleanup ConstrainedList and supplemental_semantic_id
Rename supplementary_semantic_id to supplemental_semantic_id, as it's
named in the spec.

Use the official description of supplemental_semantic_id from the spec
in the docstrings.

Only implement the bare minimum of list functions for ConstrainedList.
Other functions such as `append()`, `pop()` and `remove()` are inherited
from MutableSequence and implemented via `__getitem__()`, `__len__()`,
`insert()`, and `__delitem__()`.

Document Constraint AASd-118 in HasSemantics docstring.

Change type of `supplemental_semantic_id` in `__init__()` of classes
that inherit from `HasSemantics`.
Optional[ConstrainedList[Reference]] creates an ambiguity between None
and an empty list, which both correspond to no supplemental semantic ids
being present. Insead, initialize the set with an empty iterable.
Also, instead of a ConstrainedList[Reference], use Iterable[Reference]
as we don't want the users of this SDK to worry about ConstrainedList,
it should be an implementation detail.

Add item_set_hook to ConstrainedList. Similar to item_del_hook, it is
actually not necessary for the HasSemantics constraint, but I only
noticed that later when I re-read the HasSemantics constraint.

Instead of raising a ValueError when a constraint is violated, raise an
AASConstraintViolation, which is a special class of this SDK for exactly
this purpose.

Clean up the tests a bit, e.g. use pythons unittest assertion functions
instead of raising a ValueError on error. Also split the tests by
functionality in separate functions.

Improve the ConstrainedListTest by verifying which values are passed to
the hook functions.