model: implement constraints for constrained string type aliases
The constraints are implemented as functions in a new
`model._string_constraints` module. Each of the functions verifies
individual length and pattern constraints for the respective string
type. The check-function for the respective string alias must be
called, whenever a new value is assigned to an attribute, that is
of the type of that string alias.

This is realized via decorator-functions, that are defined for each
string alias. These decorators extend the decorated class by a property
attribute for the given attribute-name. The setter of that property
calls the check-function whenever a new value is set. All
setter-functions allow setting `None` values, because the decorator
may be used for optional attributes as well. If `None` would be assigned
to mandatory attributes, the typechecker catches this error.

Wherever an attribute is already defined as a class property, the
setter is extended by a call to the respective check-function.