Partially revert "Remove attribute `kind` from `SubmodelElement`"
This partially reverts 77c33e773b3e22365627c1f59ed2d5e09fd12ea0. The
attribute is only removed from `SubmodelElement`s in the spec, but the
commit also removes deserialization of the `kind`-Attribute for
`Submodels` and some `kind` attributes of `Qualifier`s and `Submodel`s
in the examples.