model.base: add `ConstrainedLangStringSet` and subclasses
This adds the abstract base class `ConstrainedLangStringSet` and the
following subclasses:
- `MultiLanguageNameType`
- `MultiLanguageTextType`
- `DefinitionTypeIEC61360`
- `PreferredNameTypeIEC61360`
- `ShortNameTypeIEC61360`

`ConstrainedLangStringSet` is implemented as an abstract subclass of
`LangStringSet`, which uses a given function to check the constraints
of all stored values, similar to `ConstrainedList`. This function is
called for every key-value pair on initialization and whenever a new
key-value pair is added. Additionally, functions such as `__init__()`
and `__setitem__()` of the base class `LangStringSet` are called where
appropriate, such that all constraints for language tags are also
checked here.

Each of the concrete subclasses consist of just an `__init__()`
function, that passes a constraint-check function to
`ConstrainedLangStringSet.__init__()`.