model.base: add `item_id_{set,del}_hook` to `NamespaceSet`
These hooks allow dynamically assigning a value to the identifying
attribute of an object whenever it is added to a `NamespaceSet`
(`item_id_set_hook`). Furthermore, the `item_id_del_hook` allows
unsetting the identifying attribute whenever an object is removed
from a `NamespaceSet`.
This functionality is necessary for the implementation of
`SubmodelElementLists`, because there we need to store items without
an identifying attribute in a `NamespaceSet`.