examples.data._helper: disable comparison of unordered `SubmodelElementList`
Since direct children of `SubmodelElementList` don't have an identifying
attribute anymore (AASd-120), they cannot be compared because it is
impossible to know which SubmodelElement should be compared against
which other element. Maybe this can be implemented again in the future,
when hashing is implemented for all SubmodelElements, but for now we
raise a `NotImplementedError`.

A test-case for this behavior is added and `order_relevant` is set to
`true` in all example files.