model.submodel: assign values correctly in `Entity.__init__`
The attributes need to be assigned bypassing the setters because the
setters try to access attributes that haven't been set yet for
constraint validation. Only `global_asset_id` is set via the setter
as a final constraint validation, and because `global_asset_id` is
also constrained via a string constraint.