model.submodel: verify constraints when `SpecificAssetIds` are added to `Entity`
This adds an `item_add_hook` to the `specific_asset_id`
`ConstrainedList`, which is called whenever a new item is added to the
list. This is necessary because a co-managed `Entity` is not allowed to
have specific asset ids, so it shouldn't be possible to add any in this
case.