compliance_tool: add verification of `Extensions`
Previously, the compliance tool would ignore `Extensions` of objects.
However, during the update to V3.0 and the changes to the JSON/XML
formats, we noticed that the adapters aren't properly tested this way.

The validation of `Extensions` can be disabled via the flag
`--dont-check-extensions` or the keyword argument `check_extensions` to
`AASDataChecker`, since many users are probably only interested in
comparing the main model, without any extensions.