Fix negative xsd range edge cases (#178)
The current implementation uses the `abs()` function to check
that the value is inside the range.
However, the range is not symmetric.
For example, for a [byte](https://www.w3.org/TR/xmlschema11-2/#byte),
which is between [-128, 127],
the value of
[maxInclusive](https://www.w3.org/TR/xmlschema11-2/#dt-maxInclusive)
should be 127,
and [minInclusive](https://www.w3.org/TR/xmlschema11-2/#dt-minInclusive)
 should be -128.