adapter: Refactor XML and JSON de-/serialization methods (#166)
Previously, the methods for handling attributes 
of abstract classes in XML and JSON 
de-/serialization were quite big and hard to 
understand. 

This PR refactors these methods.  Mostly a 
method extraction refactoring is done, 
to simplify the methods and to keep less 
abstraction levels in a single method.

The method `_expect_type` is renamed to 
`_validate_type` as it clearer represents 
what the method does.