Refactor dependency management to pyproject.toml
Previously, our dependency management was a bit of
a mess. We had both a `requirements.txt` and a
`pyproject.toml` and some dependencies were only
defined in one but not the other.

This aims to fix this mess and refactor towards
only using `pyproject.toml` to define the
dependencies.

At the same time, we also clean up the install
process, which now works by simply calling
`pip install .` or `pip install -e .[dev]` for the
developer envrionment. This change is reflected in
the `.github/workflows`.

Furthermore, we configure the already introduced
`setuptools_scm` to automatically infer a version
based on the most recent git tag.