http.py: Fix cursor in response and set JSON Response as default (#311)
Previously, our `cursor` in a http response was
an integer (as it is a number). The specification 
asks however for this attribute ot be a string. 
Therefore, we adapt the type. 

Fixes #309

Previously, we returned XML by default, if the 
`accept_mimetypes` had `*/*` in it. This is non-
compliant to the specification, which expects JSON
to be returned. We fix the issue to return JSON by
default now instead.

Fixes #310