app/main.py: Fix default storage_path (#7)
Previously, trying to run the container using
the tutorial from the `README.md` led to an error,
due to a wrongly set default storage path if
the environment variable `STORAGE_PATH` is not 
set. This fixes this issue by setting the default
storage path to `/storage` rather than the prior
relative path: `storage`.