Dockerfile: Change the path to the main Eclipse BaSyx Python SDK repository
Previously, we used the `rwth-iat` fork to pull the SDK from,
inside the docker file, due to the fact, that the `eclipse-basyx`
repository, did not contain the necessary branch.
Now that this is the case, it makes sense to always use the
official repository.

This also fixes the dependency issues of missing dependencies, as
they have been fixed in the main repository.

Fix #8
Fix #9