Update sdk/pyproject.toml and .gitignore to new repository structure
After moving the SDK from the root of the
repository (`.`) to `./sdk`, we adapt the
`sdk/pyproject.toml` with the necessary changes.
Namely, `setuptools-scm` needs to know where the
repository root is. While we're at it, we change
the `write_to` option which is deprecated to the
new `version_file` option.

We also adapt the `.gitignore` by ignoring the
prior ignored files now in all directories, not
just the repository root.