Adapt Dockerfile to new BaSyx-Python SDK repository structure
Previously, the Eclipse BaSyx-Python SDK contained
just the SDK module. With the recent refactoring
of the SDK GitHub repository towards a monolithic
structure, the SDK module moved into the `./sdk`
subdirectory, see [basyx-python-sdk#317].

With the new repository structure, we obviously
have to adapt the install path of the SDK in this
Dockerfile, which has been done here.

[basyx-python-sdk#317]: https://github.com/eclipse-basyx/basyx-python-sdk/pull/317