adapter: Add adapter for RDF (#308)
This adds adapters to serialize and deserialize 
RDF files, following the specification of the 
Asset Administration Shell. 

While there are still currently issues with the 
specified RDF schema (see #308), this adapter 
works to read and write AAS RDF files. 
Due to these issues, we decided to keep this
implementation on a separate branch, until we can
be sure it is not subject to big changes anymore.

---------

Co-authored-by: s-heppner <mail@s-heppner.com>