pyproject.toml: Update version for pyshacl
Previously, we used version 0.26.0 of pyshacl.
This version however had a problem with typing,
namely the `py.typed` file was in the wrong
location (see [pyshacl#267]).

This has been fixed in this new release, therefore
we bump the version number in our dependency
declaration.

This also should fix our CI Pipeline, as we both
fix the mypy issues, as well as some pycodestyle
ones.

[pyshacl#267]: https://github.com/RDFLib/pySHACL/issues/267