Update to Part 1 (Metamodel) version 3.0.1
This updates the `sdk` to version 3.0.1 of the
metamodel. Since this was a minor bugfix release,
there are only a handful of changes and only one
that actually affects our SDK (as the rest of the
changes were editorial):

Remove the pattern check from `PathType`.
Effectively, that means we remove the pattern
from the `_string_constraints.check_path_type`
method. See: [aas-core#334].

Furthermore, we update the version table in the
`README.md`.

[aas-core#334]: https://github.com/aas-core-works/aas-core-meta/pull/334