SDK: Bump required Python version to >= 3.9, Add version CI check
This bumps the `requires-python` fiels in the
`pyproject.toml` of the SDK from `">=3.8"` to
`">=3.9"`, due to the fact that Python v3.8
reached EoL on 2024-10-07.

Furthermore, to avoid future backward
compatibility issues in native Python, as well as
to ensure we're always supporting the currently
supported Python versions, we add a CI check
`check-python-versions` that uses two new scripts
in `./etc/scripts` to check:
a) our currently supported Python versions to an
   EoL database
b) that our currently supported Python versions
   defined in the CI match the ones in the
   `pyproject.toml`.

This should ensure that issues like #330 are
avoided in the future. (Note that we still need
to fix #330 though).