adapter.http: Fix reference to constant from the future
Currently, we're using `datetime.UTC`,
a constant defined in the built-in `datetime`
module.

However, this constant was only introduced in
Python version 3.11, as you can see in the
documentation:

- Does not exist in [datetime Python 3.10]
- Exists in [datetime python 3.11]

We did not catch this, as we most likely
programmed the module with Python >= 3.11 and our
CI also ran `mypy` with Python 3.12.

[datetime Python 3.10]: https://docs.python.org/3.10/library/datetime.html#constants
[datetime python 3.11]: https://docs.python.org/3.11/library/datetime.html#datetime.UTC

Fixes #330