scripts.set_copyright_year: Add a CI check to ensure the copyright is updated (#344)
Previously, there was no check if the copyright
statement on top of each file (due to the MIT
license) was up to date. 

This adapts our `update_copyright_year.sh` script
with a checking mode to determine if any file
needs adapting and then adds this check to the CI.

At the same time, this also updates all necessary
files to the year 2025.

Fixes #260
Fixes #331

---------

Co-authored-by: s-heppner <iat@s-heppner.com>