sdk/docs: Bump versions for sphinx dependencies
Currently, the sphinx dependencies in `docs/add_requirements.txt` are
outdated, causing our CI pipeline to fail.

This bumps the versions for the additional requirements for the
compilation of the documentation to their newest working state.
Since that version needs `Python >= 3.11`, we also need to bump the
version of Python used in the CI.

Fixes #385