server/README.md: Clarify Docker build context (#398)
Previously, #386 proposed narrowing the scope of the build context for
'server/Dockerfile' to '/server'. As a consequence, `/sdk`, which is
required to build the server, would no longer be accessible. Locally,
this would necessitate explicitly referencing `/sdk` as a second
build context. When using the GitHub URL of this repository as build
context instead of a local repository, this would additionally lead to
inefficient data handling. Since Docker Compose does not yet support
sparse checkouts, specifying subdirectories like `/server` as build
context still results in the full repository being cloned. Consequently,
although `/sdk` would be downloaded, it would be excluded from the
narrowed build context and thus remain unavailable. Any attempt to
include `/sdk` separately afterwards would result in the same data being
downloaded again. As this is an inefficient solution, #386 is rejected.

This documents the decision of keeping the repository root as the build
context of `server/Dockerfile`. A corresponding note was added in
`server/README.md`.

Additionally, a mention of faulty line endings being a common error when
building the Docker container has been added in `server/README.md`.

Fixes #386