sdk: Add tutorial for Submodel navigation
Previously, we lacked a tutorial demonstrating how to navigate a
Submodel's hierarchy using IdShorts and IdShortPaths.

This adds a simple tutorial for Submodel navigation and updates the
documentation to reference it. The end user is shown how to navigate
SubmodelElements, such as simple Properties, Property Collections,
Property Lists and Collection Lists.

Fixes #351

# Conflicts:
#	sdk/README.md