Refactor `UniqueIdShortNamespace.get_referable`
- Renamed id_short to id_short_path, as
id_short_path also can be a single id_short

- Before the `get_referable` could only accept
a single id_short or id_short_path in form of a list.
In this commit we added parsing of `id_short_path`,
so now it is possible to handover `id_short_path`
as a string
e.g. "SomeList[3].SomeCollection.SomeProperty"